+++ tasks/seq/frolova_e_matrix_multiplication/include/ops_seq.hpp
@@ -9,7 +9,7 @@
 namespace frolova_e_matrix_multiplication_seq {
 
 void randomNumVec(int N, std::vector<int>& vec);
-std::vector<int> Multiplication(int M, int N, int K, const std::vector<int>& A, const std::vector<int>& B);
+std::vector<int> Multiplication(size_t M, size_t N, size_t K, const std::vector<int>& A, const std::vector<int>& B);
 
 class matrixMultiplication : public ppc::core::Task {
  public:
